Refactor of woe and its metrics, from src to src.metrics
